pyproject.toml
[project]
name = "learning-spark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "delta-spark>=4.0.0",
    "graphviz>=0.21",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "itables>=2.4.4",
    "nbformat>=5.10.4",
    "plotly>=6.3.0",
    "pyspark>=4.0.0",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.1",
]

# Package paths
[tool.uv.build-backend]
module-name = "configs"
module-root = ""


[tool.uv.sources]
utils = { path = "./utils" }

# build system
[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"
